scripts/requirements.txt
google-generativeai>=0.3.0
scikit-learn>=1.0.0
numpy>=1.21.0
matplotlib>=3.5.0

